Start work on refactoring
